add numeric.py
